fix: rename folder structure
